adding styling and test files to each component. Have half of components refactored from classes.
