+++ src/app.scss
@@ -1,4 +1,4 @@
-$backgroundColor: #f5f5f5;
+$backgroundColor: lightblue;
 
 html, body {
   height: 100%;
@@ -17,14 +17,14 @@ body {
   }
 
   footer {
-    background: darken($backgroundColor, 50);
+    background: darken($backgroundColor, 30);
     border-top: 1px solid #eee;
     padding: 1em 3em;
     margin: 0;
   }
 
   header {
-    background: darken($backgroundColor, 75);
+    background: darken($backgroundColor, 30);
     border-bottom: 1px solid #eee;
     color: #eee;
     padding: 0;
