--- src/components/footer/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-class Footer extends React.Component {
-  render() {
-    return <footer>&copy; 2018</footer>;
-  }
-}
-
-export default Footer;
