+++ src/components/footer/index.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Footer = () => {
+  return(
+    <footer>&copy; 2018</footer>
+  )
+}
+
+// class Footer extends React.Component {
+//   render() {
+//     return <footer>&copy; 2018</footer>;
+//   }
+// }
+
+export default Footer;
