+++ src/components/form/form.scss
@@ -1,4 +1,4 @@
-$backgroundColor: #efefef;
+$backgroundColor: #4682B4;
 $radius: 40%;
 
 form {
@@ -13,12 +13,13 @@ form {
     &.methods {
       span {
         cursor: pointer;
+        $radius: 10px;
         margin-right: 1rem;
         padding: .25rem .5rem;
-        background: #ccc;
+        background: #4682B4;
 
         &:hover, &.active {
-          background: #fff;
+          background: #b3cee5;
         }
       }
     }
