--- src/components/header/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-class Header extends React.Component {
-  render() {
-    return (
-      <header>
-        <h1>RESTy</h1>
-      </header>
-    );
-  }
-}
-
-export default Header;
