+++ src/components/header/index.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+const Header = () => {
+  return(
+    <header>
+    <h1>RESTy</h1>
+  </header>
+  )
+}
+
+// class Header extends React.Component {
+//   render() {
+//     return (
+//       <header>
+//         <h1>RESTy</h1>
+//       </header>
+//     );
+//   }
+// }
+
+export default Header;
