+++ src/components/results/index.jsx
@@ -1,5 +1,15 @@
 import React from 'react';
 
+// const Results = () => {
+//   return(
+//     <>
+//     <section>
+//     <pre>{this.props.data ? JSON.stringify(this.props.data, undefined, 2) : null}</pre>
+//   </section>
+//   </>
+//   )
+// }
+
 class Results extends React.Component {
   render() {
     return (
